Store impact in CVSS classes
Part of #683

+++ src/main/java/com/sap/oss/phosphor/fosstars/data/owasp/model/Cvssv3.java
@@ -13,7 +13,7 @@ public class Cvssv3 implements Serializable {
   private float baseScore;
 
   @JsonProperty("baseScore")
-  public float getBaseScore() {
+  public double getBaseScore() {
     return baseScore;
   }
 }
